+++ src/main/emergencyService/SeaKingHelicopter.java
@@ -27,7 +27,7 @@ public class SeaKingHelicopter extends EmergencyService{
         return 0;
     }
 
-    public double getCapacity() {
+    public int getCapacity() {
         return 0;
     }
 }
