+++ src/main/interfaces/IMeasurement.java
@@ -8,5 +8,5 @@ package main.interfaces;
 
 public interface IMeasurement {
     double getSize();
-    double getCapacity();
+    int getCapacity();
 }
