+++ src/main/ships/CruiseShip.java
@@ -26,7 +26,8 @@ public class CruiseShip extends Ship {
         return 0;
     }
 
-    public double getCapacity() {
+    public int getCapacity() {
         return 0;
     }
+
 }
