+++ src/main/ships/SpeedBoat.java
@@ -26,7 +26,7 @@ public class SpeedBoat extends Ship {
         return 0;
     }
 
-    public double getCapacity() {
+    public int getCapacity() {
         return 0;
     }
 }
